+++ Quantcast-iOS-Measurement/QuantcastMeasurement.m
@@ -1261,11 +1261,13 @@ static void QuantcastReachabilityCallback(SCNetworkReachabilityRef target, SCNet
 }
 
 -(void)logSDKError:(NSString*)inSDKErrorType withError:(NSError*)inErrorOrNil errorParameter:(NSString*)inErrorParametOrNil {
-    if ( !self.isOptedOut && self.isMeasurementActive ) {
+    if ( !self.isOptedOut ) {
         [self launchOnQuantcastThread:^(NSDate *timestamp) {
-            QuantcastEvent* e = [QuantcastEvent logSDKError:inSDKErrorType withErrorObject:inErrorOrNil errorParameter:inErrorParametOrNil withSessionID:self.currentSessionID eventTimestamp:timestamp applicationInstallID:self.appInstallIdentifier];
-            
-            [self recordEvent:e];
+            if( self.isMeasurementActive ){
+                QuantcastEvent* e = [QuantcastEvent logSDKError:inSDKErrorType withErrorObject:inErrorOrNil errorParameter:inErrorParametOrNil withSessionID:self.currentSessionID eventTimestamp:timestamp applicationInstallID:self.appInstallIdentifier];
+                
+                [self recordEvent:e];
+            }
         }];
     }